会员列表记录-购买套餐次数

+++ src/main/java/com/xiliulou/electricity/domain/car/UserCarRentalPackageDO.java
@@ -109,6 +109,6 @@ public class UserCarRentalPackageDO {
     /**
      * 套餐购买次数
      */
-    private Long payCount;
+    private Integer payCount;
 
 }
